+++ src/main/java/mod/tjt01/sprinkle/config/SprinkleConfig.java
@@ -1,6 +1,7 @@
 package mod.tjt01.sprinkle.config;
 
 import net.minecraftforge.common.ForgeConfigSpec;
+import net.minecraftforge.fml.ModList;
 import net.minecraftforge.fml.config.ModConfig;
 import org.apache.commons.lang3.tuple.Pair;
 
@@ -24,7 +25,13 @@ public class SprinkleConfig {
     public static boolean greenDyeEnabled = true;
     public static boolean brownDyeEnabled = true;
 
+    public static boolean bundlesEnabled = true;
+
     public static Map<String, Boolean> flags = new HashMap<>();
+
+    private static boolean isModLoaded(String modName) {
+        return ModList.get().isLoaded(modName);
+    }
 
     public static void bakeCommon(ModConfig config) {
         jukeboxCapabilityEnabled = COMMON_CONFIG.jukeboxTweakEnabled.get();
@@ -35,8 +42,11 @@ public class SprinkleConfig {
         smoothNightVisionFadeTime = COMMON_CONFIG.smoothNightVisionFadeTime.get();
         smoothNightVisionEnabled = COMMON_CONFIG.smoothNightVisionEnabled.get();
 
+        bundlesEnabled = COMMON_CONFIG.bundlesEnabled.get() && (COMMON_CONFIG.forceBundles.get() || !isModLoaded("quark"));
+
         flags.put("green_dye", greenDyeEnabled);
         flags.put("brown_dye", brownDyeEnabled);
+        flags.put("bundles", bundlesEnabled);
     }
 
 }